fixed type int

+++ client/game/game_Surfaces_manager.h
@@ -45,7 +45,7 @@ class GameSurfacesManager {
     void createNecessaryNpcs(std::vector<npc_t>& npcs);
     void createNecessaryCreatures(std::vector<creature_t>& creatures);
     void createNecessaryItems(std::vector<item_t>& items);
-    void createNecessaryFrameItems(std::vector<int>& items);
+    void createNecessaryFrameItems(std::vector<uint8_t>& items);
 
 public:
     //Constructor
