+++ client/game/game_render.h
@@ -51,7 +51,7 @@ public:
     void renderCreatures(std::vector<creature_t>& creatures);
     void renderItems(std::vector<item_t>& items);
     void renderEquipped(player_t& player);
-    void renderInventory(std::vector<int>& inventory);
+    void renderInventory(std::vector<uint8_t>& inventory);
 
     //Inicializador de SDL
     int init();
